requirements.txt
flask
requests
python-dotenv
opencv-python